+++ src/tests/source/Conv2dMemFuseTest1.cc
@@ -62,7 +62,7 @@ void load_rnd_img(int x, int y, int z, int size, pdb::PDBClient &pdbClient,
 
       for (int i = 0; i < x; i++) {
         for (int j = 0; j < y; j++) {
-          double data = (bool)gen() ? distn(e2) : distp(e2);
+          float data = (bool)gen() ? distn(e2) : distp(e2);
           (*(chan->getRawDataHandle()))[i * y + j] = data;
         }
       }
@@ -181,7 +181,7 @@ void verify_data(pdb::PDBClient &pdbClient, std::string dbName,
     for (int c = 0; c < r->get_num_channels(); c++) {
       pdb::Handle<FFMatrixBlock> chan = r->getMatrixAtIndex(c);
       std::cout << "Channel " << c << std::endl;
-      double *data = chan->getValue().rawData->c_ptr();
+      float *data = chan->getValue().rawData->c_ptr();
       for (int i = 0; i < r->x * r->y; i++) {
         std::cout << data[i] << ",";
       }
@@ -225,7 +225,7 @@ void verify_matrices(pdb::PDBClient &pdbClient, std::string dbName,
   for (auto r : it) {
     std::cout << "Block " << r->getBlockRowIndex() << ","
               << r->getBlockColIndex() << std::endl;
-    double *data = r->getValue().rawData->c_ptr();
+    float *data = r->getValue().rawData->c_ptr();
     for (int i = 0; i < r->getColNums() * r->getRowNums(); i++) {
       std::cout << data[i] << ",";
     }
